Move When and Map tests to separate files
